+++ examples/tensorflow/sage/test_sage.py
@@ -89,7 +89,7 @@ def prepare_connected_caveman_graph(working_dir):
         meta_path=meta_file,
         partition_count=1,
         output_dir=working_dir.name,
-        decoder_class=decoders.JsonDecoder,
+        decoder_class=decoders.JsonDecoder(),
     ).convert()
 
     logger.info(working_dir.name)