+++ examples/tensorflow/sage/test_sage_link.py
@@ -203,7 +203,7 @@ def prepare_user_teams_graph(working_dir):
         meta_path=meta_file,
         partition_count=1,
         output_dir=working_dir,
-        decoder_class=decoders.JsonDecoder,
+        decoder_class=decoders.JsonDecoder(),
     ).convert()