+++ src/python/deepgnn/graph_engine/snark/tests/convert_test.py
@@ -148,7 +148,7 @@ def test_sanity_node_map(triangle_graph):
         meta_path=meta_name,
         partition_count=1,
         output_dir=output.name,
-        decoder_class=decoder,
+        decoder_class=decoder(),
     ).convert()
 
     with open("{}/node_{}_{}.map".format(output.name, 0, 0), "rb") as nm:
@@ -175,7 +175,7 @@ def test_sanity_node_index(triangle_graph):
         meta_path=meta_name,
         partition_count=1,
         output_dir=output.name,
-        decoder_class=decoder,
+        decoder_class=decoder(),
     ).convert()
     with open("{}/node_{}_{}.index".format(output.name, 0, 0), "rb") as ni:
         expected_size = 3 * 8 + 8
@@ -196,7 +196,7 @@ def test_sanity_node_feature_index(triangle_graph):
         meta_path=meta_name,
         partition_count=1,
         output_dir=output.name,
-        decoder_class=decoder,
+        decoder_class=decoder(),
     ).convert()
     with open("{}/node_features_{}_{}.index".format(output.name, 0, 0), "rb") as ni:
         expected_size = (
@@ -222,7 +222,7 @@ def test_sanity_neighbors_index(triangle_graph):
         meta_path=meta_name,
         partition_count=1,
         output_dir=output.name,
-        decoder_class=decoder,
+        decoder_class=decoder(),
     ).convert()
     with open("{}/neighbors_{}_{}.index".format(output.name, 0, 0), "rb") as ni:
         expected_size = 3 * 8 + 8  # 3 nodes + 8 as final close
@@ -243,7 +243,7 @@ def test_sanity_edge_index(triangle_graph):
         meta_path=meta_name,
         partition_count=1,
         output_dir=output.name,
-        decoder_class=decoder,
+        decoder_class=decoder(),
     ).convert()
     with open("{}/edge_{}_{}.index".format(output.name, 0, 0), "rb") as ei:
         expected_size = 4 * 24  # 3 nodes + last line as final close
@@ -279,7 +279,7 @@ def test_sanity_edge_features_index(triangle_graph):
         meta_path=meta_name,
         partition_count=1,
         output_dir=output.name,
-        decoder_class=decoder,
+        decoder_class=decoder(),
     ).convert()
     with open("{}/edge_features_{}_{}.index".format(output.name, 0, 0), "rb") as ni:
         expected_values = [0, 24, 24, 32, 37]
@@ -302,7 +302,7 @@ def test_sanity_edge_features_data(triangle_graph):
         meta_path=meta_name,
         partition_count=1,
         output_dir=output.name,
-        decoder_class=decoder,
+        decoder_class=decoder(),
     ).convert()
     with open("{}/edge_features_{}_{}.data".format(output.name, 0, 0), "rb") as ni:
         expected_size = 37  # last value in edge_features_index
@@ -327,7 +327,7 @@ def test_sanity_metadata(triangle_graph):
         meta_path=meta_name,
         partition_count=1,
         output_dir=output.name,
-        decoder_class=decoder,
+        decoder_class=decoder(),
     ).convert()
     with open("{}/meta.txt".format(output.name), "r") as ni:
         result = ni.readlines()
@@ -376,7 +376,7 @@ def test_edge_alias_tables(triangle_graph):
         meta_path=meta_name,
         partition_count=2,
         output_dir=output.name,
-        decoder_class=decoder,
+        decoder_class=decoder(),
         dispatcher=d,
     ).convert()
     with open("{}/edge_0_0.alias".format(output.name), "rb") as ea:
@@ -430,7 +430,7 @@ def test_node_alias_tables(triangle_graph):
         meta_path=meta_name,
         partition_count=2,
         output_dir=output.name,
-        decoder_class=decoder,
+        decoder_class=decoder(),
         dispatcher=d,
     ).convert()
 
@@ -598,7 +598,7 @@ def test_sanity_node_sparse_features_index(graph_with_sparse_features):
         meta_path=meta_name,
         partition_count=1,
         output_dir=output.name,
-        decoder_class=decoder,
+        decoder_class=decoder(),
     ).convert()
     with open("{}/node_features_{}_{}.index".format(output.name, 0, 0), "rb") as ni:
         expected_size = 88
@@ -620,7 +620,7 @@ def test_sanity_node_sparse_features_data(graph_with_sparse_features):
         meta_path=meta_name,
         partition_count=1,
         output_dir=output.name,
-        decoder_class=decoder,
+        decoder_class=decoder(),
     ).convert()
     with open("{}/node_features_{}_{}.data".format(output.name, 0, 0), "rb") as nfd:
         expected_size = 128  # last value in edge_features_index
@@ -665,7 +665,7 @@ def test_sanity_edge_sparse_features_index(graph_with_sparse_features):
         meta_path=meta_name,
         partition_count=1,
         output_dir=output.name,
-        decoder_class=decoder,
+        decoder_class=decoder(),
     ).convert()
     with open("{}/edge_features_{}_{}.index".format(output.name, 0, 0), "rb") as ei:
         expected_size = 152
@@ -688,7 +688,7 @@ def test_sanity_edge_sparse_features_data(graph_with_sparse_features):
         meta_path=meta_name,
         partition_count=1,
         output_dir=output.name,
-        decoder_class=decoder,
+        decoder_class=decoder(),
     ).convert()
     with open("{}/edge_features_{}_{}.data".format(output.name, 0, 0), "rb") as efd:
         expected_size = 394  # last value in edge_features_index